+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.20.41
 
+* `Repos`:
+    * `Exposed`:
+        * `AbstractExposedKeyValueRepo`, `ExposedKeyValueRepo`, `AbstractExposedKeyValuesRepo`, `ExposedKeyValuesRepo` got opportunity to setup some part of their flows
+
 ## 0.20.40
 
 * `Versions`: